Update Dependencies
Changes
-------
* Update ESLint to v9 (v10 can come at a later date)
* Update dependencies where possible.  Note stream-json not updated to
  v2.0 (see JSCO-71), TS kept below v6.0 and @types/node kept at 20 (see
  JSCO-72).
