+++ package.json
@@ -35,36 +35,41 @@
   "dependencies": {
     "stream-json": "^1.9.1"
   },
+  "overrides": {
+    "diff": "^9.0.0",
+    "serialize-javascript": "^7.0.5"
+  },
   "devDependencies": {
     "@arethetypeswrong/cli": "^0.18.2",
-    "@trivago/prettier-plugin-sort-imports": "^4.3.0",
-    "@types/chai": "^5.2.2",
-    "@types/debug": "^4.1.12",
+    "@eslint/js": "^9.39.2",
+    "@trivago/prettier-plugin-sort-imports": "^6.0.2",
+    "@types/chai": "^5.2.3",
+    "@types/debug": "^4.1.13",
     "@types/mocha": "^10.0.10",
     "@types/node": "^20.14.5",
     "@types/stream-json": "^1.7.8",
-    "@typescript-eslint/eslint-plugin": "^7.13.1",
-    "@typescript-eslint/parser": "^7.13.1",
-    "chai": "^4.5.0",
-    "eslint": "^8.57.0",
+    "@typescript-eslint/eslint-plugin": "^8.54.0",
+    "@typescript-eslint/parser": "^8.54.0",
+    "chai": "^6.2.2",
+    "eslint": "^9.39.2",
     "eslint-config-prettier": "^9.1.0",
-    "eslint-import-resolver-typescript": "^4.4.4",
-    "eslint-plugin-import": "^2.32.0",
-    "eslint-plugin-jsdoc": "^48.2.12",
-    "eslint-plugin-mocha": "^10.4.3",
-    "eslint-plugin-node": "^11.1.0",
-    "ini": "^5.0.0",
-    "mocha": "^10.8.2",
-    "npm-check-updates": "^16.14.20",
-    "nyc": "^17.0.0",
-    "prettier": "^3.3.2",
-    "semver": "^7.6.2",
-    "ts-mocha": "^10.0.0",
+    "eslint-plugin-jsdoc": "^50.6.2",
+    "eslint-plugin-mocha": "^10.5.0",
+    "eslint-plugin-n": "^17.17.0",
+    "globals": "^15.14.0",
+    "ini": "^7.0.0",
+    "mocha": "^11.7.5",
+    "npm-check-updates": "^22.2.0",
+    "nyc": "^18.0.0",
+    "prettier": "^3.8.3",
+    "semver": "^7.8.0",
+    "ts-mocha": "^11.1.0",
     "ts-node": "^10.9.2",
-    "tsup": "^8.5.0",
-    "typedoc": "^0.25.13",
+    "tsup": "^8.5.1",
+    "typedoc": "^0.28.19",
     "typescript": "^5.4.5",
-    "why-is-node-running": "^2.3.0"
+    "typescript-eslint": "^8.54.0",
+    "why-is-node-running": "^3.2.2"
   },
   "repository": {
     "type": "git",
